refactor(#17): move all tags entities to one file
