+++ src/core/state/notes.ts
@@ -1,16 +1,6 @@
-import { createApi, createEvent, createStore } from 'effector';
+import { createApi, createStore } from 'effector';
 
 import { INote } from '../Note';
-import { $tags } from './tags';
-
-export const $activeTag = createStore<string | null>(null);
-
-export const setActiveTag = createEvent<string | null>();
-$activeTag.on(setActiveTag, (_, activeTag) => activeTag);
-$activeTag.on($tags, (activeTag, tags) => {
-	const isActiveTagExists = tags.some(({ id }) => id === activeTag);
-	return isActiveTagExists ? activeTag : null;
-});
 
 export const $openedNotes = createStore<INote[]>([]);
 export const openedNotesControls = createApi($openedNotes, {
