+++ src/core/state/tags.ts
@@ -13,3 +13,12 @@ export const tagAttachmentChanged = createEvent<{
 	tagId: string;
 	state: 'add' | 'delete'
 }>();
+
+export const $activeTag = createStore<string | null>(null);
+export const setActiveTag = createEvent<string | null>();
+
+$activeTag.on(setActiveTag, (_, activeTag) => activeTag);
+$activeTag.on($tags, (activeTag, tags) => {
+	const isActiveTagExists = tags.some(({ id }) => id === activeTag);
+	return isActiveTagExists ? activeTag : null;
+});
